start redesign of api
